requirements.txt
markdown>=3.6
jinja2>=3.1
pyyaml>=6
feedparser>=6.0
requests>=2.31
